rename updated input
