+++ app/ui/components/ImageUpload.tsx
@@ -30,56 +30,59 @@ export function ImageUpload({ id, name }: { id: string; name: string }) {
     }
     reader.readAsDataURL(file)
   }
-
+  function handleDivClick() {
+    inputRef.current?.click()
+  }
   return (
-    <section>
-      <div className="flex items-center gap-8 self-stretch cursor-pointer">
-        <div
-          className={clsx(
-            "flex flex-col justify-center items-center bg-light-purple rounded-xl pl-[39px] pr-[38px] pt-[61px] pb-[60px]",
-            {
-              "text-white bg-dark-grey/25": isImageUpload,
-            }
-          )}
-          style={{
-            backgroundImage: imgUrl ? `url(${imgUrl})` : "none",
-            backgroundSize: "cover",
-            backgroundPosition: "center",
-          }}
-        >
-          <SlIcon
-            aria-hidden={true}
-            name="card-image"
-            className={clsx("w-10 h-10 mb-2 text-purple", {
-              "text-white": isImageUpload,
-            })}
-          />
-          <span
-            className={clsx(
-              "text-purple text-base font-semibold leading-[150%]",
-              {
-                "text-white": isImageUpload,
-              }
-            )}
-          >
-            {isImageUpload ? "Change Image" : "+ Upload Image"}
-          </span>
-        </div>
-        <div>
-          <p className="text-dark-grey text-base font-normal leading-[150%]">
-            Image {isImageUpload ? "" : "Not"} Uploaded
-          </p>
-          <input
-            aria-hidden={true}
-            className="opacity-0 h-1 w-1"
-            type="file"
-            id={id}
-            name={name}
-            accept=".jpg, .jpeg, .png"
-            onChange={handleImage}
-          ></input>
-        </div>
-      </div>
-    </section>
+    <div
+      tabIndex={0}
+      onKeyDown={key => {
+        const keysThatActivateAInput =
+          key.code === "Space" || key.code === "Enter"
+        if (!keysThatActivateAInput) {
+          return
+        }
+        key.preventDefault()
+        handleDivClick()
+      }}
+      onClick={handleDivClick}
+      className={clsx(
+        "cursor-pointer flex flex-col justify-center items-center bg-light-purple rounded-xl pl-[39px] pr-[38px] pt-[61px] pb-[60px]",
+        {
+          "text-white bg-dark-grey/25": isImageUpload,
+        }
+      )}
+      style={{
+        backgroundImage: imgUrl ? `url(${imgUrl})` : "none",
+        backgroundSize: "cover",
+        backgroundPosition: "center",
+      }}
+    >
+      <SlIcon
+        aria-hidden={true}
+        name="card-image"
+        className={clsx("w-10 h-10 mb-2 text-purple", {
+          "text-white": isImageUpload,
+        })}
+      />
+      <span
+        className={clsx("text-purple text-base font-semibold leading-[150%]", {
+          "text-white": isImageUpload,
+        })}
+      >
+        {isImageUpload ? "Change Image" : "+ Upload Image"}
+      </span>
+      <input
+        ref={inputRef}
+        aria-hidden={true}
+        className="opacity-0 h-1 w-1"
+        type="file"
+        id={id}
+        name={name}
+        accept=".jpg, .jpeg, .png"
+        onChange={handleImage}
+        tabIndex={-1}
+      ></input>
+    </div>
   )
 }
